chore: move sierra to casm compiler trait to dedicated file
