+++ crates/starknet_sierra_compile/src/lib.rs
@@ -1,20 +1,10 @@
 //! A lib for compiling Sierra into Casm.
-use cairo_lang_starknet_classes::casm_contract_class::CasmContractClass;
-use cairo_lang_starknet_classes::contract_class::ContractClass;
-
-use crate::errors::CompilationUtilError;
 
 pub mod cairo_lang_compiler;
 pub mod config;
 pub mod errors;
+pub mod sierra_to_casm_compiler;
 pub mod utils;
 
 #[cfg(test)]
 pub mod test_utils;
-
-pub trait SierraToCasmCompiler: Send + Sync {
-    fn compile(
-        &self,
-        contract_class: ContractClass,
-    ) -> Result<CasmContractClass, CompilationUtilError>;
-}
