+++ crates/starknet_sierra_compile/src/sierra_to_casm_compiler.rs
@@ -1,0 +1,11 @@
+use cairo_lang_starknet_classes::casm_contract_class::CasmContractClass;
+use cairo_lang_starknet_classes::contract_class::ContractClass;
+
+use crate::errors::CompilationUtilError;
+
+pub trait SierraToCasmCompiler: Send + Sync {
+    fn compile(
+        &self,
+        contract_class: ContractClass,
+    ) -> Result<CasmContractClass, CompilationUtilError>;
+}
